prepare rendering more than just one action type
